fet: create commands unit
